Refactor email handling by introducing community support service, enhancing AI response generation, and improving email thread management. Updated email utilities to handle allowed recipients and CC emails, and streamlined conversation message processing.
